try cuasm&nvbit

+++ chapter3/add.cu
@@ -3,36 +3,36 @@
 #include <stdio.h>
 
 
-const double EPSILON = 1.0e-15;
-const double a = 1.23;
-const double b = 2.34;
-const double c = 3.57;
+const float EPSILON = 1.0e-15;
+const float a = 1.23;
+const float b = 2.34;
+const float c = 3.57;
 
 
 // 核函数。
-__global__ void add(const double *x, const double *y, double *z, const int N);
+__global__ void add(const float *x, const float *y, float *z, const int N);
 
 // 重载设备函数。
-__device__ double add_in_device(const double x, const double y);
-__device__ void add_in_device(const double x, const double y, double &z);
+__device__ float add_in_device(const float x, const float y);
+__device__ void add_in_device(const float x, const float y, float &z);
 
 // 主机函数。
-void check(const double *z, const int N);
+void check(const float *z, const int N);
 
 
 int main()
 {
     const int N = 1e8;
-    const int M = sizeof(double) * N;
+    const int M = sizeof(float) * N;
 
     // 申请主机内存。
-    // double *h_x = (double*) malloc(M);
-    // double *h_y = (double*) malloc(M);
-    // double *h_z = (double*) malloc(M);
+    // float *h_x = (float*) malloc(M);
+    // float *h_y = (float*) malloc(M);
+    // float *h_z = (float*) malloc(M);
     // 支持使用 new-delete 方式创建和释放内存。
-    double *h_x = new double[N];
-    double *h_y = new double[N];
-    double *h_z = new double[N];
+    float *h_x = new float[N];
+    float *h_y = new float[N];
+    float *h_z = new float[N];
 
     // 初始化主机数据。
     for (int i = 0; i < N; ++i)
@@ -43,7 +43,7 @@ int main()
 
     // 申请设备内存。
     // cudeError_t cudaMalloc(void **address, size_t size);
-    double *d_x, *d_y, *d_z;
+    float *d_x, *d_y, *d_z;
     cudaMalloc((void**)&d_x, M);
     cudaMalloc((void**)&d_y, M);
     cudaMalloc((void**)&d_z, M);
@@ -81,14 +81,14 @@ int main()
     return 0;
 }
 
-// __global__ void add(const double *x, const double *y, double *z, const int N)
+// __global__ void add(const float *x, const float *y, float *z, const int N)
 // {
 //     const int n = blockDim.x * blockIdx.x + threadIdx.x;
 //     if (n >= N) return;
 //     z[n] = x[n] + y[n];
 // }
 
-__global__ void add(const double *x, const double *y, double *z, const int N)
+__global__ void add(const float *x, const float *y, float *z, const int N)
 {
     // 在主机函数中需要依次对每个元素进行操作，需要使用一个循环。
     // 在设备函数中，因为采用“单指令-多线程”方式，所以可以去掉循环、只要将数组元素索引和线程索引一一对应即可。
@@ -106,17 +106,17 @@ __global__ void add(const double *x, const double *y, double *z, const int N)
     }
 }
 
-__device__ double add_in_device(const double x, const double y)
+__device__ float add_in_device(const float x, const float y)
 {
     return x + y;
 }
 
-__device__ void add_in_device(const double x, const double y, double &z)
+__device__ void add_in_device(const float x, const float y, float &z)
 {
     z = x + y;
 }
 
-void check(const double *z, const int N)
+void check(const float *z, const int N)
 {
     bool has_error = false;
     for (int i = 0; i < N ;++i)
